Added signup redirect url parameter
This works the same as the signin redirect. If a user was trying
to access a page and gets redirected to sign in or sign up,
after succeeding they will be redirected from where they came.

Since a user can only be redirected to either the signin or signup
page, it is beyond the scope of this feature to pass the query
parameter in the url to the other page when requested.

Updated signup redirect to work like signin

Added signup redirect for SIGNIN_AFTER_SIGNUP